+++ sdk/spring/azure-spring-boot/CHANGELOG.md
@@ -16,6 +16,9 @@
 - Deprecated azure-spring-boot-metrics-starter.
 - Change group id from `com.microsoft.azure` to `com.azure.spring`.
 
+### New Features
+- Support acquiring access tokens of previously consented permissions for web-hosted resources from Azure AD.
+
 ## 2.3.5 (2020-09-14)
 ### Key Bug Fixes
 - Get full list of groups the user belongs to from Graph API